+++ neuropsydia/path.py
@@ -5,8 +5,8 @@ class Path:
     def root():
         return(inspect.getfile(Path).split("path.py")[0])
     def logo():
-        return(inspect.getfile(Path).split("path.py")[0] + "files\\logo\\")
+        return(inspect.getfile(Path).split("path.py")[0] + r"files/logo/")
     def binary():
-        return(inspect.getfile(Path).split("path.py")[0] + "files\\binary\\")
+        return(inspect.getfile(Path).split("path.py")[0] + r"files/binary/")
     def font():
-        return(inspect.getfile(Path).split("path.py")[0] + "files\\font\\")
+        return(inspect.getfile(Path).split("path.py")[0] + r"files/font/")