+++ neuropsydia/signal.py
@@ -1,8 +1,14 @@
 # -*- coding: utf-8 -*-
-import mne
 import pandas as pd
 import numpy as np
+import datetime
 
+from .stats import z_score  # process_EDA()
+from .miscellaneous import get_creation_date  # acq_to_df()
+
+import cvxopt as cv  # process_EDA()
+import cvxopt.solvers  # process_EDA()
+import bioread  # acq_to_df()
 
 # ==============================================================================
 # ==============================================================================
@@ -12,22 +18,121 @@ import numpy as np
 # ==============================================================================
 # ==============================================================================
 # ==============================================================================
-def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
-           solver=None, options={'reltol':1e-9}):
+def acq_to_df(file, sampling_rate=1000, method="mean"):
     """
-    CVXEDA Convex optimization approach to electrodermal activity processing
-
-    This function implements the cvxEDA algorithm described in "cvxEDA: a
-    Convex Optimization Approach to Electrodermal Activity Processing"
-    (http://dx.doi.org/10.1109/TBME.2015.2474131, also available from the
-    authors' homepages).
+    Format a BIOPAC's AcqKnowledge file into a pandas' dataframe.
 
     Parameters
     ----------
-       y
-           observed EDA signal (we recommend normalizing it: y = zscore(y))
-       delta
-           sampling interval (in seconds) of y
+    file =  str
+        the path of a BIOPAC's AcqKnowledge file
+    sampling_rate = int
+        final sampling rate (samples/second)
+    method = str
+        "mean" or "pad", resampling method
+
+    Returns
+    ----------
+    df = pandas.DataFrame()
+        the dataframe
+
+
+    Example
+    ----------
+    >>> import neuropsydia as n
+    >>> n.start(False)
+    >>>
+    >>> df = acq_to_df('file.acq')
+
+    Authors
+    ----------
+    Dominique Makowski
+
+    Dependencies
+    ----------
+    - pandas
+    - bioread
+    - datetime
+    """
+    # Read file
+    creation_date = get_creation_date(file)
+    file = bioread.read(file)
+
+    # Convert creation date
+    creation_date = datetime.datetime.fromtimestamp(creation_date)
+
+
+    # Get the channel frequencies
+    freq_list = []
+    for channel in file.named_channels:
+        freq_list.append(file.named_channels[channel].samples_per_second)
+
+    # Get data with max frequency and the others
+    data = {}
+    data_else = {}
+    for channel in file.named_channels:
+        if file.named_channels[channel].samples_per_second == max(freq_list):
+            data[channel] = file.named_channels[channel].data
+        else:
+            data_else[channel] = file.named_channels[channel].data
+
+    # Create index
+    time = []
+    beginning_date = creation_date - datetime.timedelta(0, max(file.time_index))
+    for timestamps in file.time_index:
+        time.append(beginning_date + datetime.timedelta(0, timestamps))
+
+    df = pd.DataFrame(data, index=time)
+
+    # Create resampling factor
+    sampling_rate = str(int(1000/sampling_rate)) + "L"
+
+
+    # max frequency must be 1000
+    for channel in data_else:
+        channel_frequency = file.named_channels[channel].samples_per_second
+        serie = data_else[channel]
+        index = list(np.arange(0, max(file.time_index), 1/channel_frequency))
+        index = index[:len(serie)]
+        # Create index
+        time = []
+        for timestamps in index:
+            time.append(beginning_date + datetime.timedelta(0, timestamps))
+        data_else[channel] = pd.Series(serie, index=time)
+    df2 = pd.DataFrame(data_else)
+
+
+    # Resample
+    if method == "mean":
+        df2 = df2.resample(sampling_rate).mean()
+        df = df.resample(sampling_rate).mean()
+    if method == "pad":
+        df2 = df2.resample(sampling_rate).pad()
+        df = df.resample(sampling_rate).pad()
+    df = pd.concat([df, df2], 1)
+
+    return(df)
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+def process_EDA(EDA_raw, sampling_rate, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2, solver=None, options={'reltol':1e-9}):
+    """
+    A convex optimization approach to electrodermal activity processing (CVXEDA)
+
+    This function implements the cvxEDA algorithm described in "cvxEDA: a
+    Convex Optimization Approach to Electrodermal Activity Processing" (Greco et al., 2015).
+
+    Parameters
+    ----------
+       EDA_raw
+           observed EDA signal (we recommend normalizing it: EDA_raw = zscore(EDA_raw))
+       sampling_rate
+           sampling rate (samples/seconds) of EDA_raw
        tau0
            slow time constant of the Bateman function
        tau1
@@ -45,12 +150,12 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
 
     Returns
     ----------
-       r
+       phasic
            phasic component
+       tonic
+           tonic component
        p
            sparse SMNA driver of phasic component
-       t
-           tonic component
        l
            coefficients of tonic spline
        d
@@ -76,17 +181,18 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
     - cvxopt
     - numpy
     """
-    import cvxopt as cv
-    import cvxopt.solvers
+    frequency = 1/(sampling_rate/100)
 
-    n = len(y)
-    y = cv.matrix(y)
+    EDA_raw = z_score(EDA_raw)
+
+    n = len(EDA_raw)
+    EDA_raw = cv.matrix(EDA_raw)
 
     # bateman ARMA model
     a1 = 1./min(tau1, tau0) # a1 > a0
     a0 = 1./max(tau1, tau0)
-    ar = np.array([(a1*delta + 2.) * (a0*delta + 2.), 2.*a1*a0*delta**2 - 8.,
-        (a1*delta - 2.) * (a0*delta - 2.)]) / ((a1 - a0) * delta**2)
+    ar = np.array([(a1*frequency + 2.) * (a0*frequency + 2.), 2.*a1*a0*frequency**2 - 8.,
+        (a1*frequency - 2.) * (a0*frequency - 2.)]) / ((a1 - a0) * frequency**2)
     ma = np.array([1., 2., 1.])
 
     # matrices for ARMA model
@@ -95,7 +201,7 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
     M = cv.spmatrix(np.tile(ma, (n-2,1)), np.c_[i,i,i], np.c_[i,i-1,i-2], (n,n))
 
     # spline
-    delta_knot_s = int(round(delta_knot / delta))
+    delta_knot_s = int(round(delta_knot / frequency))
     spl = np.r_[np.arange(1.,delta_knot_s), np.arange(delta_knot_s, 0., -1.)] # order 1
     spl = np.convolve(spl, spl, 'full')
     spl /= max(spl)
@@ -112,7 +218,7 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
     nC = C.size[1]
 
     # Solve the problem:
-    # .5*(M*q + B*l + C*d - y)^2 + alpha*sum(A,1)*p + .5*gamma*l'*l
+    # .5*(M*q + B*l + C*d - EDA_raw)^2 + alpha*sum(A,1)*p + .5*gamma*l'*l
     # s.t. A*q >= 0
 
     old_options = cv.solvers.options.copy()
@@ -124,7 +230,7 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
         G = cv.sparse([[-A,z(2,n),M,z(nB+2,n)],[z(n+2,nC),C,z(nB+2,nC)],
                     [z(n,1),-1,1,z(n+nB+2,1)],[z(2*n+2,1),-1,1,z(nB,1)],
                     [z(n+2,nB),B,z(2,nB),cv.spmatrix(1.0, range(nB), range(nB))]])
-        h = cv.matrix([z(n,1),.5,.5,y,.5,.5,z(nB,1)])
+        h = cv.matrix([z(n,1),.5,.5,EDA_raw,.5,.5,z(nB,1)])
         c = cv.matrix([(cv.matrix(alpha, (1,n)) * A).T,z(nC,1),1,gamma,z(nB,1)])
         res = cv.solvers.conelp(c, G, h, dims={'l':n,'q':[n+2,nB+2],'s':[]})
         obj = res['primal objective']
@@ -133,10 +239,10 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
         Mt, Ct, Bt = M.T, C.T, B.T
         H = cv.sparse([[Mt*M, Ct*M, Bt*M], [Mt*C, Ct*C, Bt*C],
                     [Mt*B, Ct*B, Bt*B+gamma*cv.spmatrix(1.0, range(nB), range(nB))]])
-        f = cv.matrix([(cv.matrix(alpha, (1,n)) * A).T - Mt*y,  -(Ct*y), -(Bt*y)])
+        f = cv.matrix([(cv.matrix(alpha, (1,n)) * A).T - Mt*EDA_raw,  -(Ct*EDA_raw), -(Bt*EDA_raw)])
         res = cv.solvers.qp(H, f, cv.spmatrix(-A.V, A.I, A.J, (n,len(f))),
                             cv.matrix(0., (n,1)), solver=solver)
-        obj = res['primal objective'] + .5 * (y.T * y)
+        obj = res['primal objective'] + .5 * (EDA_raw.T * EDA_raw)
     cv.solvers.options.clear()
     cv.solvers.options.update(old_options)
 
@@ -146,9 +252,11 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
     q = res['x'][:n]
     p = A * q
     r = M * q
-    e = y - r - t
+    e = EDA_raw - r - t
 
-    return(np.array(a).ravel() for a in (r, p, t, l, d, e, obj))
+    results = (np.array(a).ravel() for a in (r, t, p, l, d, e, obj))
+
+    return(results)
 
 
 # ==============================================================================
@@ -161,7 +269,7 @@ def cvxEDA(y, delta, tau0=2., tau1=0.7, delta_knot=10., alpha=0.4, gamma=1e-2,
 # ==============================================================================
 def extract_peak(channel_data, value="max", size=0):
     """
-    Exctract the peak (max or min) of one or several channels.
+    Extract the peak (max or min) of one or several channels.
 
     Parameters
     ----------
@@ -193,7 +301,6 @@ def extract_peak(channel_data, value="max", size=0):
 
     Dependencies
     ----------
-    - mne > 0.13.0
     - numpy
     - pandas
     """
@@ -226,33 +333,30 @@ def extract_peak(channel_data, value="max", size=0):
 # ==============================================================================
 # ==============================================================================
 # ==============================================================================
-def triggers_from_photodiode(photo_channel, names=None, treshold=0.04):
+def binarize_signal(signal, treshold, upper=True):
     """
-    Create MNE compatible triggers based on a photodiode channel.
+    Binarize a channel based on a continuous channel.
 
     Parameters
     ----------
-    photo_channel = MNE channel
-        The photodiode channel.
-    names = list
-        A list of event names.
+    signal = array or list
+        The signal channel.
     treshold = float
-        The treshold to select the triggers.
+        The treshold.
+    upper = bool
+        Associate a 1 with a value above or under the treshold.
 
     Returns
     ----------
-    tuple
-        (events, event_id)
+    list
+        binary_signal
 
     Example
     ----------
     >>> import neuropsydia as n
     >>> n.start(False)
     >>>
-    >>> raw = mne.io.read_raw("eeg_file")
-    >>> photo_channel = raw.copy().pick_channels(['PHOTO'])
-    >>> events, event_id = triggers_from_photodiode(photo_channel)
-    >>> raw.add_events(events, stim_channel="STI 014")
+    >>> binary_signal = binarize_signal(signal, treshold=4)
     >>>
     >>> n.close()
 
@@ -262,47 +366,217 @@ def triggers_from_photodiode(photo_channel, names=None, treshold=0.04):
 
     Dependencies
     ----------
-    - mne > 0.13.0
-    - numpy
+    None
     """
-    original_names = names.copy()
+    signal = list(signal)
+    binary_signal = []
+    for i in range(len(signal)):
+        if upper == True:
+            if signal[i] > treshold:
+                binary_signal.append(1)
+            else:
+                binary_signal.append(0)
+        if upper == False:
+            if signal[i] < treshold:
+                binary_signal.append(1)
+            else:
+                binary_signal.append(0)
+    return(binary_signal)
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
 
-    if names != None:
-        event_names = list(set(names))
-        event_index = [1, 2, 3, 4, 5, 32]
-        event_id = {}
-        for i in enumerate(event_names):
-            names = [event_index[i[0]] if x==i[1] else x for x in names]
-            event_id[i[1]] = event_index[i[0]]
+def events_onset(signal, treshold, upper=True, time_index=None):
+    """
+    Find the onsets of all events based on a continuous signal.
 
-    # Extract data from one channel
-    data, times = photo_channel[:]
-    T = list(data.T)
+    Parameters
+    ----------
+    signal = array or list
+        The signal channel.
+    treshold = float
+        The treshold.
+    upper = bool
+        Associate a 1 with a value above or under the treshold.
+    time_index = array or list
+        Add a corresponding datetime index, will return an addional array with the onsets as datetimes.
 
-    events = []
-    for i in range(len(times)):
-        if T[i] < treshold:
-            events.append(1)
-        else:
-            events.append(0)
+    Returns
+    ----------
+    list or tuple of lists
+        events onsets
 
-    event_times = []
-    for i in range(len(events)):
+    Example
+    ----------
+    >>> import neuropsydia as n
+    >>> n.start(False)
+    >>>
+    >>> events_onset = events_onset(signal, treshold=4)
+    >>>
+    >>> n.close()
+
+    Authors
+    ----------
+    Dominique Makowski
+
+    Dependencies
+    ----------
+    None
+    """
+    binary_data = binarize_signal(signal, treshold=treshold, upper=upper)
+
+    events_time = []
+    events_onset = []
+    for i in range(len(binary_data)):
         if i > 0:
-            if events[i]==1 and events[i-1]==0:
-                event_times.append(i)
-
-    if len(event_times) != len(original_names):
-        print("NEUROPSYDIA ERROR: triggers_from_photodiode(): length of trigger names vector does not match the number of detected triggers (n = " +
-              str(len(event_times)) + "), change names or crop the raw data")
-    if names != None:
-        events = np.array([event_times, [0]*len(event_times), names]).T
-        return(events, event_id, names)
+            if binary_data[i]==1 and binary_data[i-1]==0:
+                events_onset.append(i)
+                if time_index is not None:
+                    events_time.append(time_index[i])
+    if time_index is None:
+        return(events_onset)
     else:
-        events = np.array([event_times, [0]*len(event_times), [1]*len(event_times)]).T
-        return(events, event_id)
+        return(events_onset, events_time)
 
-    
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+
+def create_epochs(signal, events_onset, sampling_rate, onset=-250, duration=1000, stimuli_name=None):
+    """
+    Create a dataframe of epoched data.
+
+    Parameters
+    ----------
+    signal = array or list
+        the signal.
+    events_onset = list
+        list of events onsets (see `events_onset()` function).
+    sampling_rate = int
+        Signal's sampling rate (samples/second).
+
+    Returns
+    ----------
+    pandas' dataframe
+        timepoints * epochs
+
+    Example
+    ----------
+    >>> import neuropsydia as n
+    >>> n.start(False)
+    >>>
+    >>> epochs = create_epochs(signal)
+    >>>
+    >>> n.close()
+
+    Authors
+    ----------
+    Dominique Makowski
+
+    Dependencies
+    ----------
+    - numpy
+    - pandas
+    """
+    length = int(sampling_rate * duration / 1000)
+    length_onset =  int(sampling_rate * abs(onset) / 1000)
+
+    events_onset = np.array(events_onset)
+    events_onset = events_onset + onset
 
 
+    epoch = 0
+    data = {}
+    for i in range(len(signal)):
+        try:
+            if i == events_onset[epoch]:
+                data[epoch] = []
+                j = 0
+                for j in range(length):
+                    data[epoch].append(signal[i+j])
+                epoch += 1
+        except IndexError:
+            pass
+
+    index = np.array(list(range(length)))-length_onset
+    epochs = pd.DataFrame.from_dict(data)
+    epochs.index = index
+    if stimuli_name is not None:
+        if len(list(set(stimuli_name))) != len(stimuli_name):
+            print("Neuropsydia error: create_epochs(): stimuli names must be all different. See create_evoked()")
+        else:
+            epochs.columns = stimuli_name
+    return(epochs)
+
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+# ==============================================================================
+
+def create_evoked(epochs, events, average=True):
+    """
+    Create a dictionary containing evoked data.
+
+    Parameters
+    ----------
+    epochs = dataframe
+        epoched data.
+    events = list
+        list of events types.
+    average = bool
+        Set True for collapsing the epoched.
+
+    Returns
+    ----------
+    dic
+        A dictionary containing one dataframe for each event.
+
+    Example
+    ----------
+    >>> import neuropsydia as n
+    >>> n.start(False)
+    >>>
+    >>> events = ["emotion", "neutral", "emotion", "neutral"]
+    >>> evoked = create_evoked(epochs)
+    >>>
+    >>> n.close()
+
+    Authors
+    ----------
+    Dominique Makowski
+
+    Dependencies
+    ----------
+    - numpy
+    - pandas
+    """
+    index = epochs.index
+
+    evoked = {}
+    for event in set(events):
+        evoked[event] = {}
+
+    for event in enumerate(events):
+        evoked[event[1]][event[0]] = epochs[event[0]]
+
+    for event in evoked:
+        evoked[event] = pd.DataFrame.from_dict(evoked[event])
+        evoked[event].index = index
+        if average == True:
+            evoked[event] = evoked[event].mean(axis=1)
+    return(evoked)